[change] refactor data container to another file
